refactor: split module into multiple files
