+++ async/_common.ts
@@ -1,0 +1,10 @@
+import type { AsyncOperator } from "./operator.ts";
+
+/**
+ * @internal
+ */
+export type LastAsyncOperatorReturn<
+  T extends AsyncOperator<unknown, unknown>[],
+> = T extends
+  [...AsyncOperator<unknown, unknown>[], AsyncOperator<unknown, infer R>] ? R
+  : never;
