+++ operator.ts
@@ -1,0 +1,4 @@
+/**
+ * Operator function that takes a value and returns a new value.
+ */
+export type Operator<A, B> = (v: A) => B;
